Move sorting and grouping into list logic
